+++ CHANGELOG.md
@@ -24,4 +24,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Restructuring pipeline by [@chaimain](https://github.com/chaimain).
 - Work on various Models functions and assignments by [@chaimain](https://github.com/chaimain).
 - Adding plotting functions and flux-points step by [@chaimain](https://github.com/chaimain).
--
+- Start using time intervals and light-curve analysis step by [@chaimain](https://github.com/chaimain).
+-